Started replacing double with float

+++ include/wembed.h
@@ -51,64 +51,64 @@ struct Edge {
 struct TimingResult {
     uint64_t depth;
     std::string displayName;
-    double value;
+    float value;
 };
 
 // Loss values from the most recent embedding step
 struct Loss {
-    double attractive;
-    double repulsive;
-    double total;
+    float attractive;
+    float repulsive;
+    float total;
 };
 
 struct Options {
     // Embedding parameters
     int32_t embeddingDimension = 4;
     bool useUnitWeights = false;                      // true: degree-based weights; false: unit weights
-    double dimensionHint = -1.0;                 // hint for the dimension of the input graph (-1 = auto)
+    float dimensionHint = -1.0;                 // hint for the dimension of the input graph (-1 = auto)
     bool layeredEmbedding = false;               // multilevel embedding via graph coarsening
 
     // Force parameters
     SpatialIndex indexType = IndexSprk;
-    double attractionScale = 1.0;
-    double repulsionScale = 1.0;
-    double centreScale = 0.0;                    // pull toward origin; nonzero enables it (useful for unconnected graphs)
-    double edgeLength = 1.0;
-    double expansionStretch = 1.0;               // stretch applied during layer expansion
+    float attractionScale = 1.0;
+    float repulsionScale = 1.0;
+    float centreScale = 0.0;                    // pull toward origin; nonzero enables it (useful for unconnected graphs)
+    float edgeLength = 1.0;
+    float expansionStretch = 1.0;               // stretch applied during layer expansion
 
     // Gradient descent parameters
     OptimizerType optimizerType = OptimizerAdam;
     int32_t maxIterations = 10000;
-    double simpleOptMaxDisplacement = 1.0;       // per-step cap (only used when optimizerType == OptimizerSimple)
+    float simpleOptMaxDisplacement = 1.0;       // per-step cap (only used when optimizerType == OptimizerSimple)
 
     // Learning rate schedule (lr* prefix). Every parameter states which schedules read it.
     LRSchedule lrSchedule = LRExponentialCooling;
-    double learningRate = 10.0;                  // initial learning rate (both schedules)
+    float learningRate = 10.0;                  // initial learning rate (both schedules)
     int32_t warmupSteps = 20;                    // linear LR ramp-up over the first steps (both schedules)
-    double lrCoolingFactor = 0.995;              // per-step multiplicative decay, lower = faster cooldown (LRExponentialCooling only)
-    double lrDecayFactor = 0.5;                  // multiplicative drop on a decay event (LRLossAdaptive only)
-    double lrDecayThreshold = 1e-2;              // decay when the loss-decrease rate stays below this (LRLossAdaptive only)
+    float lrCoolingFactor = 0.995;              // per-step multiplicative decay, lower = faster cooldown (LRExponentialCooling only)
+    float lrDecayFactor = 0.5;                  // multiplicative drop on a decay event (LRLossAdaptive only)
+    float lrDecayThreshold = 1e-2;              // decay when the loss-decrease rate stays below this (LRLossAdaptive only)
     int32_t lrAdaptPatience = 20;                // consecutive in-zone steps before a decay OR growth event (LRLossAdaptive only)
-    double lrGrowthFactor = 1.0;                 // multiplicative growth while the loss keeps decreasing fast
+    float lrGrowthFactor = 1.0;                 // multiplicative growth while the loss keeps decreasing fast
                                                  // (LRLossAdaptive only; 1.0 disables growth -> pure plateau decay)
-    double lrGrowthThreshold = 1e-1;             // grow when the loss-decrease rate stays above this (LRLossAdaptive only)
+    float lrGrowthThreshold = 1e-1;             // grow when the loss-decrease rate stays above this (LRLossAdaptive only)
 
     // Stopping criterion (maxIterations always applies as a hard cap).
     StopCriterion stopCriterion = StopLoss;  // which signal terminates the run
 
     // Displacement stopping criterion (StopDisplacement).
-    double stopDisplacementTol = 3e-4;           // relative per-step node movement (mean displacement / radius of
+    float stopDisplacementTol = 3e-4;           // relative per-step node movement (mean displacement / radius of
                                                  // gyration) below which the layout counts as settled
     int32_t stopDisplacementPatience = 5;        // settled steps in a row before stopping
 
     // Shared loss-progress signal (loss* prefix): windowed relative loss decrease rate(t),
     // read by both the loss stop criterion and the LRLossAdaptive schedule.
-    double lossSmoothingFactor = 0.3;            // EMA weight of the newest loss sample before the monitor sees it
+    float lossSmoothingFactor = 0.3;            // EMA weight of the newest loss sample before the monitor sees it
                                                  // (1.0 disables smoothing); a light denoise on rate(t)
     int32_t lossRateWindow = 30;                 // steps over which the relative loss-decrease rate is measured
 
     // Loss stagnation stopping criterion (StopLoss).
-    double stopLossTol = 1e-3;                   // ftol: converged once rate(t) stays below this (relative decrease over window)
+    float stopLossTol = 1e-3;                   // ftol: converged once rate(t) stays below this (relative decrease over window)
     int32_t stopLossPatience = 50;               // sub-tolerance steps in a row before stopping
 };
 
@@ -166,14 +166,14 @@ class Embedder {
     int32_t getNumVertices() const;
     int32_t getEmbeddingDimension() const;
 
-    // flat copy of coordinates getNumVertices() * getEmbeddingDimension() doubles, row-major.
-    void copyCoordinatesTo(double* out) const;
+    // flat copy of coordinates getNumVertices() * getEmbeddingDimension() floats, row-major.
+    void copyCoordinatesTo(float* out) const;
 
     Graph getCurrentGraph() const;
-    std::vector<std::vector<double>> getCoordinates() const;
-    std::vector<double> getWeights() const;
-    void setCoordinates(const std::vector<std::vector<double>>& coordinates);
-    void setWeights(const std::vector<double>& weights);
+    std::vector<std::vector<float>> getCoordinates() const;
+    std::vector<float> getWeights() const;
+    void setCoordinates(const std::vector<std::vector<float>>& coordinates);
+    void setWeights(const std::vector<float>& weights);
 
     // Hierarchical breakdown of time spent in each phase of the embedding.
     std::vector<TimingResult> getTimings() const;
@@ -183,15 +183,15 @@ class Embedder {
 
     // Learning rate the optimizer used in the most recent step
     // (before the first step: the initial learning rate).
-    double getCurrentLearningRate() const;
+    float getCurrentLearningRate() const;
 
     // Relative node displacement of the most recent step
     // (mean per-node movement / radius of gyration); the displacement stop watches this.
-    double getLastRelDisplacement() const;
+    float getLastRelDisplacement() const;
 
     // Windowed relative loss-decrease rate of the most recent step (rate(t));
     // the loss stop watches this (a step is stagnant when it stays below stopLossTol).
-    double getLastRelLossImprovement() const;
+    float getLastRelLossImprovement() const;
 
     void writeCoordinates(const std::string& filePath, bool writeWeights = true) const;
 
@@ -211,7 +211,7 @@ Graph graphFromEdgeListFile(const std::string& filePath,
 
 // Read a coordinate file (one row per vertex). Useful for resuming an embedding
 // via Embedder::setCoordinates.
-std::vector<std::vector<double>> readCoordinatesFromFile(
+std::vector<std::vector<float>> readCoordinatesFromFile(
     const std::string& filePath,
     const std::string& comment = "%",
     const std::string& delimiter = ",");
