+++ src/embeddingLib/include/embedder/ConvergenceMonitor.hpp
@@ -14,33 +14,33 @@
  */
 class ConvergenceMonitor {
    public:
-    static constexpr double STILL_IMPROVING = std::numeric_limits<double>::infinity();
+    static constexpr float STILL_IMPROVING = std::numeric_limits<float>::infinity();
 
-    ConvergenceMonitor(double relTol, int patience, double smoothingFactor, int rateWindow);
+    ConvergenceMonitor(float relTol, int patience, float smoothingFactor, int rateWindow);
 
-    void observe(double loss);
+    void observe(float loss);
 
     bool converged() const { return numStagnantSteps >= patience; }
     int stagnantSteps() const { return numStagnantSteps; }
-    double relImprovement() const { return lastRate; }  // rate(t); STILL_IMPROVING during warmup
+    float relImprovement() const { return lastRate; }  // rate(t); STILL_IMPROVING during warmup
     int numObservations() const { return numObserved; }
-    double lastLoss() const { return lastObservedLoss; }
+    float lastLoss() const { return lastObservedLoss; }
 
    private:
-    static constexpr double TINY = 1e-12;
+    static constexpr float TINY = 1e-12;
 
-    double relTol;
+    float relTol;
     int patience;
-    double smoothingFactor;
+    float smoothingFactor;
     int rateWindow;
 
-    std::vector<double> ring;  // last rateWindow + 1 smoothed losses
+    std::vector<float> ring;  // last rateWindow + 1 smoothed losses
     int ringHead = 0;          // next write slot == oldest retained sample once full
     int ringCount = 0;
 
-    double smoothedLoss = 0.0;
+    float smoothedLoss = 0.0;
     int numObserved = 0;
-    double lastObservedLoss = 0.0;
-    double lastRate = STILL_IMPROVING;
+    float lastObservedLoss = 0.0;
+    float lastRate = STILL_IMPROVING;
     int numStagnantSteps = 0;
 };
