+++ src/embeddingLib/include/embeddingSpace/DotProduct.hpp
@@ -5,8 +5,8 @@
 
 class DotProduct : public Embedding {
    public:
-    DotProduct(const std::vector<std::vector<double>> &coords);
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    DotProduct(const std::vector<std::vector<float>> &coords);
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
 
    private:
