+++ src/embeddingLib/include/embeddingSpace/InfNorm.hpp
@@ -5,8 +5,8 @@
 
 class InfNorm : public Embedding {
    public:
-    InfNorm(const std::vector<std::vector<double>> &coords);
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    InfNorm(const std::vector<std::vector<float>> &coords);
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
 
    private:
