+++ src/embeddingLib/include/embeddingSpace/MercatorEmbedding.hpp
@@ -5,18 +5,18 @@
 
 class MercatorEmbedding : public Embedding {
    public:
-    MercatorEmbedding(const std::vector<double>& radii, const std::vector<std::vector<double>>& positions);
-    MercatorEmbedding(const std::vector<double>& radii, const std::vector<double>& thetas);
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    MercatorEmbedding(const std::vector<float>& radii, const std::vector<std::vector<float>>& positions);
+    MercatorEmbedding(const std::vector<float>& radii, const std::vector<float>& thetas);
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
 
    private:
     const int DIMENSION;
     VecList coordinates;
-    std::vector<double> thetas;
-    std::vector<double> radii;
+    std::vector<float> thetas;
+    std::vector<float> radii;
 
-    double S1_distance(double r1, double r2, double theta1, double theta2) const;
-    double compute_angle_d_vectors(CVecRef v1, CVecRef v2) const;
-    double SD_distance(double r1, double r2, CVecRef pos1, CVecRef pos2) const;
+    float S1_distance(float r1, float r2, float theta1, float theta2) const;
+    float compute_angle_d_vectors(CVecRef v1, CVecRef v2) const;
+    float SD_distance(float r1, float r2, CVecRef pos1, CVecRef pos2) const;
 };
