+++ src/embeddingLib/include/embeddingSpace/Poincare.hpp
@@ -5,8 +5,8 @@
 
 class Poincare : public Embedding {
    public:
-    Poincare(const std::vector<std::vector<double>> &coords);
-    virtual double getSimilarity(NodeId a, NodeId b) const;
+    Poincare(const std::vector<std::vector<float>> &coords);
+    virtual float getSimilarity(NodeId a, NodeId b) const;
     virtual int getDimension() const;
 
    private:
